refactor(emitter): split §19 over-cap shards into under-2000-line siblings
Two declaration-emitter files touched by the previous emit-parity fix
exceeded the 2000-line hard cap (§19). Split each into named siblings:

- imports_and_modules.rs (2873 → 1503 lines) + parameters_and_heritage.rs (1382 lines)
- js_exports.rs (4301 → 1344 lines) + js_exports_local.rs (1649 lines) + js_exports_namespace.rs (1357 lines)

All five shards compile clean; the 3 private helpers that cross shard
boundaries are promoted to pub(super) / pub(in crate::declaration_emitter).
No behavior change — pure mechanical split. 2965 unit tests pass.
